External examples - adding support for external examples.
The idea is that writing <example> in the actual code is very clumsy and messy.
We'd rather have external files that hold the examples, and can even be run separately, without using any ngdocs compiler.

We incorporate those external examples using a hint in the documentation itself:

@external-example:some.example

This will look for a directory named 'example' inside another directory named 'some', under the base examples directory (which can be configured in the Gruntfile, and is named 'examples' by default).
Every file under that directory will become a <file> in the documentation.
The AngularJS module will have the name of the example, i.e. 'some.example'.

HTML files that are used this way can be real, full HTMLs.
The plugin will know to only take the <body> part and include it in the <file> section.